Create a blank actor nodelist data frame

The nodelist is a data frame that contains information for the actors in a simulation, including names and EPA dictionaries and equations to use. This function creates a blank nodelist. Actor lines can be added with add_actor() (recommended) or manually.

Usage

blank_nodelist(
  use.alphas = FALSE,
  use.betas = FALSE,
  use.deltas = FALSE,
  use.numsamples = FALSE
)

Arguments

use.alphas, use.betas, use.deltas

logical indicating whether to include columns for alpha, beta, or delta parameters. All default to FALSE. See the BayesACT C documentation for more information on these parameters.

use.numsamples

logical indicating whether to include a numsamples column. See the BayesACT C documentation for more information on this parameter.

Value

Empty nodelist data frame with correct column names

Examples

blank_nodelist()
#> [1] name       dict       dict_stat  dict_group eqns       eqns_group
#> <0 rows> (or 0-length row.names)
blank_nodelist(use.alphas = TRUE, use.numsamples = TRUE)
#> [1] name       dict       dict_stat  dict_group eqns       eqns_group alphas    
#> [8] numsamples
#> <0 rows> (or 0-length row.names)
blank_nodelist(use.alphas = TRUE, use.betas = TRUE, use.deltas = TRUE)
#> [1] name       dict       dict_stat  dict_group eqns       eqns_group alphas    
#> [8] betas      deltas    
#> <0 rows> (or 0-length row.names)