extract_dict_list

Handles dict input from nodelist tibbles. This can be:

a list containing one tibble with four rows containing four strings (if actdata keys or filepaths)

or a list of four tibbles representing dictionaries (if a df was passed)

A mix of these two formats

Usage

extract_dict_list(input)

Arguments

input

the input passed from the nodelist df

Value

a list with four string entries or a tibble with four nested tibble entries

Details

These are not checked for validity here; that happens later. TODO: Think about whether this check should be moved up (possibly to the nodelist creation).