Write agent into input

Compile all necessary information for an individual agent and add it to a template object that can be written to a text file

Usage

agent(
  name,
  bayesact_dir,
  dict = "usfullsurveyor2015",
  dict_stat = "mean",
  dict_gender = "average",
  eqns = "us2010",
  eqns_gender = c("av", "female"),
  opt_args = ""
)

Arguments

name

string agent name

bayesact_dir

top level directory at which the bayesact code lives

dict

string or length 4 vector, key(s) of an available dictionary OR filepaths OR data frames. If there are any file paths or data frames, it must be a length 4 vector.

dict_stat

string or length 4 vector ("mean", "sd", or "cov")

dict_gender

string or length 4 vector ("average", "male", or "female")

eqns

string or length 2 vector, key(s) of an available equation set OR filepath(s). First is gender for impression ABO equation, second is gender for emotion equation.

eqns_gender

string or length 2 vector ("av", "f", or "m"). First is gender for impression ABO equation, second is gender for emotion equation.

opt_args

more agent-level parameters that may or may not be included

Value

file with inserted lines